requirements.txt
django == 4.0.4
pillow
django-crispy-forms
django-rest-framework